+++ dtaianomaly/workflow/workflow_from_config.py
@@ -323,6 +323,9 @@ def detector_entry(entry):
     elif detector_type == "CopulaBasedOutlierDetector":
         return anomaly_detection.CopulaBasedOutlierDetector(**entry_without_type)
 
+    elif detector_type == "KShapeAnomalyDetector":
+        return anomaly_detection.KShapeAnomalyDetector(**entry_without_type)
+
     else:
         raise ValueError(f"Invalid detector entry: {entry}")